Add stores alias, move stores outside data
