+++ src/components/Auditor/AuditorImport.svelte
@@ -6,7 +6,7 @@
 
 <script>
   import { getContext } from 'svelte';
-  import { importAssertions } from 'data/stores/earl/assertionStore.js';
+  import { importAssertions } from 'stores/earl/assertionStore.js';
   import File, { readFile } from 'components/formcomponents/File.svelte';
 
   const { translate } = getContext('app');
