+++ src/stores/scopeStore.js
